moved user guides to config
